Add Git-Common and use functions by sourcing it.

+++ Git-DeleteMergedBranches.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
 
-# Switch to master branch before deleting merged branches.
-git checkout master
-git pull
+source ./Git-Common.sh
+IsGitRepository
 
-# Get all the merged branches with master.
-mergedBranches=$(git branch --merged master | grep feature*)
+if [[ $? -eq 1 ]]
+then 
+    # Switch to master branch before deleting merged branches.
+    git checkout master
+    git pull
 
-# Loop through the merged branches and delete them.
-for branch in $mergedBranches
-do 
-    git push --delete origin $branch
-    git branch --delete $branch
-done
+    # Get all the merged branches with master.
+    mergedBranches=$(git branch --merged master | grep feature*)
+
+    # Loop through the merged branches and delete them.
+    for branch in $mergedBranches
+    do 
+        git push --delete origin $branch
+        git branch --delete $branch
+    done 
+fi
